+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Deprecated: `-e` and `--extract` CLI flags, and the `extractStyleTagsFromHtml` node API option. If you use these flags or option, please consider creating a processor for the community. See the [release planning](/docs/user-guide/release-planning.md) document for more details.
 - Added: `ignoreProperties: []` option for `declaration-block-no-duplicate-properties`.
 - Added: `no-empty-source` rule.
+- Fixed: accuracy of warning positions for empty blocks when using SugarSS parser.
 
 # 6.7.1